smart blind stick
Objective
The main objective is to help the visually impaired people to detect obstacles from a farther
distance and to track the location of the visually impaired person when an unexpected situation
occurs. The technology used in this project is based on IoT integrated with GPS and GSM. It
also aims at reducing the cost to provide optimal solution to the visually impaired.